requirements.txt
alembic==1.4.2
aniso8601==8.0.0
arrow==0.15.8
attrs==21.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.6.0
dnspython==1.16.0
elasticsearch==7.7.1
email-validator==1.0.5
eventlet==0.25.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Seeder==1.1.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.1
Flask-Validator==1.4.1
greenlet==0.4.16
grpcio==1.30.0
idna==2.9
importlib-metadata==4.0.1
inflection==0.5.1
isbnlib==3.9.10
iso3166==1.0.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.22.3
monotonic==1.5
mysqlclient==1.4.6
numpy==1.18.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pandas==1.0.4
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.12.1
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2019.3
PyYAML==5.4.1
pyzbar==0.1.8
qrcode==6.1
razorpay==1.2.0
requests==2.23.0
rsa==4.0
schwifty==2020.2.2
short-url==1.2.2
six==1.14.0
SQLAlchemy==1.3.17
swagger-ui-bundle==0.0.6
typing-extensions==3.10.0.0
urllib3==1.25.9
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.4.1
zope.event==4.4
zope.interface==5.1.0
